requirements.txt

Click==7.0
Flask==1.0.3
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.4
pandas==0.24.2
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
Werkzeug==0.15.4
python-decouple==3.3
requests==2.20.1